:sparkles: add test for valid identifier fct. move to utils
